% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adface}
\alias{adface}
\title{Reactogenicity Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 51 columns.
}
\source{
Template adface associated with the \code{adsl}, \code{face}, \code{ex} and \code{vs} datasets.
}
\usage{
adface
}
\description{
A Reactogenicity analysis dataset
}
\seealso{
Other dataset: 
\code{\link{admiralvaccine_adsl}},
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}},
\code{\link{vx_adsl}}
}
\concept{dataset}
\keyword{dataset}
