% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralvaccine_adce}
\alias{admiralvaccine_adce}
\title{Clinical Events Analysis Dataset - Vaccine Specific}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 56 columns.
}
\source{
(\url{https://github.com/pharmaverse/admiralvaccine/blob/main/inst/templates/ad_adce.R})
}
\usage{
admiralvaccine_adce
}
\description{
An example Clinical Events analysis dataset
}
\seealso{
Other dataset:
\code{\link{admiralvaccine_adface}},
\code{\link{admiralvaccine_adis}},
\code{\link{admiralvaccine_adsl}}
}
\concept{dataset}
\keyword{dataset}
