% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralvaccine_adsl}
\alias{admiralvaccine_adsl}
\title{Subject Level Analysis Dataset - Vaccine Specific}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 46 columns.
}
\source{
(\url{https://github.com/pharmaverse/admiralvaccine/blob/main/inst/templates/ad_adsl.R})
}
\usage{
admiralvaccine_adsl
}
\description{
An example Subject Level analysis dataset
}
\seealso{
Other dataset:
\code{\link{admiralvaccine_adce}},
\code{\link{admiralvaccine_adface}},
\code{\link{admiralvaccine_adis}}
}
\concept{dataset}
\keyword{dataset}
