% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae}
\alias{ae}
\title{Supplemental AE Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 16 columns.
}
\usage{
ae
}
\description{
A SDTM AE dataset
}
\seealso{
Other dataset: 
\code{\link{adsl}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}},
\code{\link{vx_adsl}}
}
\concept{dataset}
\keyword{dataset}
