% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce}
\alias{ce}
\title{CE Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 31 columns.
}
\usage{
ce
}
\description{
A SDTM CE dataset
}
\seealso{
Other dataset: 
\code{\link{admiralvaccine_adsl}},
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}},
\code{\link{vx_adsl}}
}
\concept{dataset}
\keyword{dataset}
