% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_diam_to_sev.R
\name{derive_param_diam_to_sev}
\alias{derive_param_diam_to_sev}
\title{Creating Severity Records from Diameters}
\usage{
derive_param_diam_to_sev(
  dataset = NULL,
  filter_diam = "DIAMETER",
  filter_faobj = c("REDNESS", "SWELLING"),
  testcd_sev = "SEV",
  test_sev = "Severity/Intensity",
  none = c(0, 2),
  mild = c(2, 5),
  mod = c(5, 10),
  sev = 10
)
}
\arguments{
\item{dataset}{Input data set

The variables \code{USUBJID}, \code{APTPTREF}, \code{FAOBJ}, \code{FASCAT},\code{AVAL}, \code{AVALC},
\code{FAOBJ}, \code{FATESTCD} and \code{FATEST} are expected for Input data set.(\verb{data set})}

\item{filter_diam}{Diameter record filter

\emph{Default: "DIAMETER"}
\emph{Permitted Value}: A character scalar.

Helps to filter the diameter records to derive the severity records by
passing the \code{FATESTCD} value for diameter  which is corresponding to the
specified events in \code{filter_faobj}.}

\item{filter_faobj}{Event filter

\emph{Default: c("REDNESS","SWELLING")}
\emph{Permitted Value}: A character vector or Scalar.

Helps to filter the events (Redness and swelling) which has diameter records
to derive severity records by passing the events from \code{FAOBJ}.}

\item{testcd_sev}{\code{FATESTCD}value for severity

\emph{Default: "SEV"}
\emph{Permitted Value}: A Character scalar

Assign the value for \code{FATESTCD} variable to indicate the severity records.
Ignore the argument if you want to set the default value.

\emph{Note}: This argument value will be used to check whether input data set has
Severity records for specified \code{filter_faobj} event. If it has, those records
will be removed and new severity records will be derived from diameters.}

\item{test_sev}{\code{FATEST} value for severity

\emph{Default: "Severity"}
\emph{Permitted Value}: A Character scalar

Assign the value for \code{FATEST} variable to indicate the severity records.
Ignore the argument if you want to set the default value.

\emph{Assign the Severity limits to create Severity grade(AVALC)}}

\item{none}{limit for grade \code{"NONE"}

\emph{Default: c(0,2)}
\emph{Permitted value}: A Numeric vector

The \code{none} and the following arguments(\code{mild}, \code{mode} and \code{sev} ) will be
used for assigning the diameter limits to derive the \code{AVALC}(severity grade).

\emph{Note: Use the limit reference to pass the values in argument}
\emph{Since the condition was coded like this,}
\emph{NONE : none < AVAL <= none}
\emph{MILD : mild < AVAL <= mild}
\emph{MODERATE: mod < AVAL <= mod}
\emph{SEVERE: sev > AVAL}
\emph{User should pass the values as numeric vectors. Refer the default values.}}

\item{mild}{limit for grade \code{"MILD"}

\emph{Default: c(2,5)}
\emph{Permitted value}: A Numeric vector}

\item{mod}{limit for grade \code{"MODERATE"}

\emph{Default: c(5,10)}
\emph{Permitted value}: A Numeric vector}

\item{sev}{limit for grade \code{"SEVERE"}

\emph{Default: 10}
\emph{Permitted value}: A Numeric vector}
}
\value{
The Input data with the new severity records for Redness and swelling which
is specified in \code{filter_faobj} and AVAL, AVALC will be derived and fatestcd,
fatest will be changed as per the values
}
\description{
To derive the severity records from the diameter records per subject per
event per period.
}
\details{
\enumerate{
\item Pass the Input data set in \code{dataset} with required variables and \code{AVAL}
should have the diameter values for the events(\code{FAOBJ}) specified in
\code{filter_faobj}.
\item If the SDTM data set has severity record for redness and swelling, User
can skip this function to keep the SDTM level severity records for further
derivation.
\item If User want to derive the severity records from the diameter records even
that data set has severity records for Redness and swelling. This function
will remove the existing severity records and it will create the new
severity records as\code{AVALC} which is derived from diameter. and \code{AVAL} will
be derived as numeric severity grade to get the maximum severity records
which will be derived by \code{derive_param_maxsev.R}.
\item Pass the values in severity grade arguments(\code{none}, \code{mild}, \code{mod}, \code{sev})
as per the study needs.
5.User can pass the events which has diameter values to derive their severity
records by passing the appropriate \code{filter_diam}, \code{filter_faobj}and severity
grade limits.
}
}
\examples{
library(dplyr)
library(admiral)
library(tibble)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~AVAL, ~AVALC, ~ATPTREF, ~FATEST, ~FATESTCD,
  "XYZ1001", "REDNESS", 7.5, "7.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 3.5, "3.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 2, "2", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 1.8, "1.8", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "REDNESS", 1.4, "1.4", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 11.1, "11.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 7.4, "7.4", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 6, "6", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 2.1, "2.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "REDNESS", 1.1, "1.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 5.5, "5.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 2.5, "2.5", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 2, "2", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 1.8, "1.8", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1001", "SWELLING", 1.4, "1.4", "VACCINATION 1", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 10.1, "10.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 7.1, "7.1", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 5, "5", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 1.8, "1.8", "VACCINATION 2", "Diameter", "DIAMETER",
  "XYZ1002", "SWELLING", 1.4, "1.4", "VACCINATION 2", "Diameter", "DIAMETER"
)

derive_param_diam_to_sev(
  dataset = input,
  filter_faobj = c("REDENSS", "SWELLING"),
  filter_diam = "DIAMETER",
  testcd_sev = "SEV",
  test_sev = "Severity"
)

}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\concept{der_adxx}
\keyword{der_adxx}
