% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_maxdiam.R
\name{derive_param_maxdiam}
\alias{derive_param_maxdiam}
\title{Adds records for the maximum diameter values within each by group to the
output dataset.}
\usage{
derive_param_maxdiam(
  dataset,
  filter = NULL,
  by_vars = NULL,
  test_maxdiam = "Maximum Diameter",
  testcd_maxdiam = "MAXDIAM"
)
}
\arguments{
\item{dataset}{Input dataset

The variables \code{USUBJID}, \code{APTPTREF}, \code{FAOBJ}, \code{FASCAT}, \code{AVAL},
\code{FATESTCD} and \code{FATEST} are required for the input data set.(\code{dataset})}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\code{FATESTCD} and \code{FAOBJ} related to diameter records are expected to be
filtered in this step}

\item{by_vars}{Grouping variables

\emph{Default: exprs(USUBJID, FAOBJ, ATPTREF)}

Variables to consider for generation of groupwise maximum diameter records.
Providing the names of variables in \code{exprs()} will create a groupwise
summary and generate maximum diam records for the specified groups.}

\item{test_maxdiam}{Value for \code{FATEST}

\emph{Default}: "Maximum Diameter"
\emph{Permitted Value: A Character scalar}

Pass a value for \code{FATEST} which will be indicate the maximum diameter
records}

\item{testcd_maxdiam}{Value for \code{FATESTCD}

\emph{Default}: "MAXDIAM"
\emph{Permitted Value: A Character scalar}

Pass a value for \code{FATESTCD} which will be indicate the maximum diameter
records.}
}
\value{
A data frame with derived records for maximum diameter appended
to the original dataset.
}
\description{
Creates maximum diameter record per subject per event per period.
Maximum diameter records will be derived for the Administration and systemic
events which has diameter records.
}
\details{
The maximum aval value of diameter within each group is filtered.

if more than one record have same maximum values within by group then
this function will take the first record from that group.
}
\examples{
library(tibble)
library(admiral)
library(dplyr)
library(rlang)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~ATPTREF, ~AVAL, ~AVALC, ~FATEST, ~FATESTCD, ~FASCAT,
  "1", "REDNESS", "VAC1", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC1", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC1", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC2", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC2", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "1", "REDNESS", "VAC2", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC1", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC1", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC1", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC2", .5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC2", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE",
  "2", "SWELLING", "VAC2", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE"
)

# Derive maximum diameter records for redness and swelling.

derive_param_maxdiam(
  dataset = input,
  filter = FAOBJ \%in\% c("REDNESS", "SWELLING") & FATESTCD == "DIAMETER",
  by_vars = exprs(USUBJID, FAOBJ, FALNKGRP),
  test_maxdiam = "Maximum Diameter",
  testcd_maxdiam = "MAXDIAM"
)
}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\author{
Vikram S
}
\concept{der_adxx}
\keyword{der_adxx}
