% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_maxtemp.R
\name{derive_param_maxtemp}
\alias{derive_param_maxtemp}
\title{Deriving maximum temperature record}
\usage{
derive_param_maxtemp(
  dataset = NULL,
  filter_faobj = "FEVER",
  by_vars = vars(USUBJID, FAOBJ, ATPTREF),
  test_maxtemp = "Maximum temperatue",
  testcd_maxtemp = "MAXTEMP"
)
}
\arguments{
\item{dataset}{Input data set

Input data set should have the temperature record which will be derived by
using the function \code{derive_param_fever_occur}.}

\item{filter_faobj}{filter for fever record

\emph{Default: "FEVER"}
\emph{Permitted Value: A Character Scalar}

Helps to filter the fever record by passing the value from \code{FATESTCD} which
indicates the fever records.}

\item{by_vars}{Grouping variables

\emph{Default: vars(USUBJID,FAOBJ,ATPTREF)}
\emph{Permitted Value: A Character vector}

Pass the \code{by_vars} in \code{vars()} based on the by_vars variables the maximum
temperature will be derived from \code{VSSTRESN}.}

\item{test_maxtemp}{Value for \code{FATEST}

\emph{Default}: "Maximum temperature"
\emph{Permitted Value: A Character scalar}

Pass a value for \code{FATEST} which will be indicate the maximum temperature
records}

\item{testcd_maxtemp}{\emph{Default}: "MAXTEMP"
\emph{Permitted Value: A Character scalar}

Pass a value for \code{FATESTCD} which will be indicate the maximum temperature
records. refer the default value.}
}
\value{
The input data set with new records \code{FATESTCD = MAXTEMP}indicating the
maximum temperature records based on the specified variables in \code{by_vars}.
\code{DTYPE} will be populated as \code{MAXIMUM}. \code{FATEST}will be populated as
specified in \code{test_maxtemp}and \code{FATESCD} will be populated as specified in
\code{testcd_maxtemp}. Maximum temperature values will be populated in \code{AVAL}.
}
\description{
Create a summary record for maximum temperature per subject per period.
}
\examples{

derive_param_maxtemp(
dataset = input,
filter_faobj = "FEVER",
test_maxtemp = "Maximum Temperature",
testcd_maxtemp = "MAXTEMP"
)

}
