% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm}
\alias{dm}
\title{DM Dataset}
\format{
An object of class \code{data.frame} with 3 rows and 28 columns.
}
\usage{
dm
}
\description{
A SDTM DM dataset
}
\seealso{
Other dataset: 
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}}
}
\concept{dataset}
\keyword{dataset}
