% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{face}
\alias{face}
\title{FACE Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 29 columns.
}
\usage{
face
}
\description{
A SDTM FACE dataset
}
\seealso{
Other dataset: 
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}}
}
\concept{dataset}
\keyword{dataset}
