% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Derive Extension Example}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Happy Message
}
\description{
Says hello admiral
}
\details{
Something Something
}
\examples{
hello_admiral(hw = FALSE)
}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()}
}
\author{
Open Source
}
\concept{der_adxx}
\keyword{der_adxx}
