% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{is}
\alias{is}
\title{IS Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 24 columns.
}
\usage{
is
}
\description{
A SDTM IS dataset
}
\seealso{
Other dataset: 
\code{\link{suppis}}
}
\concept{dataset}
\keyword{dataset}
