% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppis}
\alias{suppis}
\title{Supplemental IS Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 10 columns.
}
\usage{
suppis
}
\description{
A SDTM Supplemental IS dataset
}
\seealso{
Other dataset: 
\code{\link{admiralvaccine_adsl}},
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{vs}},
\code{\link{vx_adsl}}
}
\concept{dataset}
\keyword{dataset}
