% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralvaccine.R
\name{tweak_rdrr_url}
\alias{tweak_rdrr_url}
\title{Tweak yml url}
\usage{
tweak_rdrr_url(...)
}
\arguments{
\item{...}{Arguments from pkgdown tweak_page}
}
\description{
Tweak yml url
}
\keyword{internal}
