% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs}
\alias{vs}
\title{VS Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 29 columns.
}
\usage{
vs
}
\description{
A SDTM VS dataset
}
\seealso{
Other dataset: 
\code{\link{admiralvaccine_adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}}
}
\concept{dataset}
\keyword{dataset}
