% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_adsl}
\alias{vx_adsl}
\title{Vx - Subject Level Analysis Dataset}
\format{
An object of class \code{data.frame} with 1 rows and 45 columns.
}
\usage{
vx_adsl
}
\description{
A subject level analysis dataset
}
\seealso{
Other dataset: 
\code{\link{adface}},
\code{\link{vx_ae}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppis}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
