% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_ae}
\alias{vx_ae}
\title{AE Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 16 columns.
}
\usage{
vx_ae
}
\description{
A SDTM AE dataset
}
\seealso{
Other dataset: 
\code{\link{adface}},
\code{\link{vx_adsl}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppis}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
