% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_suppce}
\alias{vx_suppce}
\title{Supplemental Clinical Events}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 9 columns.
}
\usage{
vx_suppce
}
\description{
A SDTM Supplementary CE dataset
}
\concept{dataset
#' @source
(\url{https://github.com/pharmaverse/admiralvaccine/tree/main/inst/create_vx_data/vx_suppis.R})}
\keyword{dataset}
