% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_suppdm}
\alias{vx_suppdm}
\title{Supplemental Demographics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 9 columns.
}
\usage{
vx_suppdm
}
\description{
A SDTM Supplementary DM dataset
}
\concept{dataset
#' @source
(\url{https://github.com/pharmaverse/admiralvaccine/tree/main/inst/create_vx_data/vx_suppdm.R})}
\keyword{dataset}
