% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_suppex}
\alias{vx_suppex}
\title{Supplemental Exposure
A SDTM Supplementary EX dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 9 columns.
}
\usage{
vx_suppex
}
\description{
Supplemental Exposure
A SDTM Supplementary EX dataset
}
\seealso{
Other dataset
#' @source
(\url{https://github.com/pharmaverse/admiralvaccine/tree/218_data_source\%40main/inst/create_vx_data/vx_suppex.R}): 
\code{\link{vx_suppface}}
}
\concept{dataset
#' @source
(\url{https://github.com/pharmaverse/admiralvaccine/tree/218_data_source\%40main/inst/create_vx_data/vx_suppex.R})}
\keyword{dataset}
