% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_suppface}
\alias{vx_suppface}
\title{Supplemental Findings About Clinical Events}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 9 columns.
}
\usage{
vx_suppface
}
\description{
A SDTM Supplementary FACE dataset
}
\seealso{
Other dataset
#' @source
(\url{https://github.com/pharmaverse/admiralvaccine/tree/main/inst/create_vx_data/vx_suppex.R}): 
\code{\link{vx_suppex}}
}
\concept{dataset
#' @source
(\url{https://github.com/pharmaverse/admiralvaccine/tree/main/inst/create_vx_data/vx_suppex.R})}
\keyword{dataset}
