% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_suppis}
\alias{vx_suppis}
\title{Supplemental IS Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 10 columns.
}
\usage{
vx_suppis
}
\description{
A SDTM Supplemental IS dataset
}
\seealso{
Other dataset: 
\code{\link{adface}},
\code{\link{vx_adsl}},
\code{\link{vx_ae}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
