% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_vs}
\alias{vx_vs}
\title{VS Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 29 columns.
}
\usage{
vx_vs
}
\description{
A SDTM VS dataset
}
\seealso{
Other dataset: 
\code{\link{adface}},
\code{\link{vx_adsl}},
\code{\link{vx_ae}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppis}}
}
\concept{dataset}
\keyword{dataset}
