% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adface}
\alias{adface}
\title{Reactogenicity Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41 rows and 51 columns.
}
\source{
Template adface associated with the \code{adsl}, \code{face}, \code{ex} and \code{vs} datasets.
}
\usage{
adface
}
\description{
A Reactogenicity analysis dataset
}
\seealso{
Other dataset: 
\code{\link{vx_adsl}},
\code{\link{vx_ae}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppis}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
