% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralvaccine.R
\docType{package}
\name{admiralvaccine-package}
\alias{admiralvaccine}
\alias{admiralvaccine-package}
\title{admiralvaccine: ADaM in R Asset Library - Extension}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralvaccine}
  \item \url{https://github.com/pharmaverse/admiralvaccine}
}

}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Vikram S
  \item Arjun Rubalingam
  \item Dhivya Kanagaraj
  \item Federico Baratin
  \item Yamini Purna Bollu
  \item Ilse Augustyns
  \item Kalyani Bodicherla
  \item Hilde Delanghe
}

Other contributors:
\itemize{
  \item Neetu Sangari [contributor]
  \item Jayashree V [contributor]
  \item Jagadish Katam [contributor]
  \item Andrea Pammolli [contributor]
  \item Daniele Bottigliengo [contributor]
  \item Ranya Ben Hsain [contributor]
  \item Marleen Nijs [contributor]
  \item Mandy Peng [contributor]
  \item Tina Zhai [contributor]
}

}
\concept{internal}
\keyword{internal}
