% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralvaccine_adface}
\alias{admiralvaccine_adface}
\title{Findings About Clinical Events Analysis Dataset - Vaccine Specific}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 371 rows and 60 columns.
}
\source{
(\url{https://github.com/pharmaverse/admiralvaccine/blob/main/inst/templates/ad_adface.R})
}
\usage{
admiralvaccine_adface
}
\description{
An example Findings About Clinical Events analysis dataset
}
\seealso{
Other dataset: 
\code{\link{admiralvaccine_adce}},
\code{\link{admiralvaccine_adis}},
\code{\link{admiralvaccine_adsl}}
}
\concept{dataset}
\keyword{dataset}
