% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralvaccine_adis}
\alias{admiralvaccine_adis}
\title{Immunogenicity Specimen Assessments Analysis Dataset - Vaccine Specific}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 102 columns.
}
\source{
(\url{https://github.com/pharmaverse/admiralvaccine/blob/main/inst/templates/ad_adis.R})
}
\usage{
admiralvaccine_adis
}
\description{
An example Immunogenicity Specimen Assessments analysis dataset
}
\seealso{
Other dataset: 
\code{\link{admiralvaccine_adce}},
\code{\link{admiralvaccine_adface}},
\code{\link{admiralvaccine_adsl}}
}
\concept{dataset}
\keyword{dataset}
