% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralvaccine_adsl}
\alias{admiralvaccine_adsl}
\title{Subject Level Analysis Dataset}
\format{
An object of class \code{data.frame} with 3 rows and 10 columns.
}
\source{
Template adsl dataset associated with the \code{dm} and \code{ex} datasets
}
\usage{
admiralvaccine_adsl
}
\description{
An example subject level analysis dataset
}
\seealso{
Other dataset: 
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}}
}
\concept{dataset}
\keyword{dataset}
