% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_fever_occur.R
\name{derive_param_fever_occur}
\alias{derive_param_fever_occur}
\title{derive_param_fever_occur.R}
\usage{
derive_param_fever_occur(dataset, source_data, faobj)
}
\arguments{
\item{dataset}{Input dataset
Input dataset is expected to have variables \code{USUBJID},\code{FAOBJ},
\code{FACAT},\code{FATESTCD} and \code{FATEST}}

\item{source_data}{Input SDTM dataset(VS)
Input SDTM dataset is expected to have
\code{VSCAT = "REACTOGENICITY" & VSTESTCD = "TEMP"}records
\emph{Default: "VS"}}

\item{faobj}{FAOBJ Value for FEVER records in output dataset
\emph{Default: "FEVER"}}
}
\value{
The output dataset contains records with \code{FATESTCD = "OCCUR"}for
\code{FAOBJ = FEVER} records.
}
\description{
Getting \code{FAOBJ = "FEVER"} records from \code{SDTM.VS} dataset, if there are no
FEVER records in input dataset
}
\details{
Check if \code{FAOBJ = FEVER} record is present in input dataset,
if not then use \code{SDTM.VS} to get FEVER records.
With temperature values from VSSTRESN we decide if FEVER
has occurred or not(\code{FAORRES = "Y"/"N"}).
Since records are derived, these FEVER records are considered
\code{DTYPE = "DERIVED"}
}
\examples{
library(tibble)
library(dplyr)
library(admiraldev)
library(admiral)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~FATESTCD, ~FACAT, ~FASCAT, ~FATPT,
  "ABC101", "REDNESS", "SEV", "REACTOGENICITY", "ADMINISTRATIVE SITE", "DAY 1",
  "ABC101", "REDNESS", "DIAM", "REACTOGENICITY", "ADMINISTRATIVE SITE", "DAY 2",
  "ABC101", "VOMITTING", "SEV", "REACTOGENICITY", "SYSTEMIC", "DAY 1",
  "ABC101", "FATIQUE", "OCCUR", "REACTOGENICITY", "SYSTEMIC", "DAY 3"
)

vs <- tribble(
  ~USUBJID, ~VSTESTCD, ~VSCAT, ~VSSTRESN, ~VSSTRESU, ~VSTPT,
  "ABC101", "TEMP", "REACTOGENICITY", 38.3, "C", "DAY 1",
  "ABC101", "TEMP", "REACTOGENICITY", 38, "C", "DAY 2",
  "ABC101", "TEMP", "REACTOGENICITY", 36, "C", "DAY 3",
  "ABC101", "TEMP", "REACTOGENICITY", 37, "C", "DAY 4",
  "ABC101", "TEMP", "REACTOGENICITY", 39, "C", "DAY 5",
  "ABC101", "TEMP", "REACTOGENICITY", 39, "C", "DAY 6",
  "ABC101", "TEMP", "REACTOGENICITY", 38, "C", "DAY 7"
)

derive_param_fever_occur(
  dataset = input,
  source_data = vs,
  faobj = "FEVER"
)

}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\author{
Dhivya Kanagaraj
}
\concept{der_adxx}
\keyword{der_adxx}
