% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_maxsev_draft.R
\name{derive_param_maxsev}
\alias{derive_param_maxsev}
\title{Derive Maximum Severity}
\usage{
derive_param_maxsev(dataset = NULL, testcd_sev = "SEV", by_vars = NULL)
}
\arguments{
\item{dataset}{The Input data frame for this function
required variables for Input data freame :USUBJID, ATPTREF(FATPTREF),
PARCAT2(FASCAT), AVAL(FASTRESN),AVALC(FASTRESC),FAOBJ,FATESTCD,FATEST.}

\item{testcd_sev}{In testcd_sev, Pass the testcd in the quotes ("SEV") which is representing
the severity records.if you have more than one testcd, pass it as a list in
the testcd_sev arugument.}

\item{by_vars}{In by_vars, pass the inside vars() grouping variables to derive the Maximum severity
eg: by_vars= vars(USUBJID,FAOBJ, ATPTREF)}
}
\value{
Dataframe which has Maxsev records (with DTYPE as "MAXIMUM", Derived AVAL
and AVALC and some grouping vars and updated FATESTCD and FATEST.)
}
\description{
Derive the maximum severity for each administrative and Systemic events.
}
\details{
Pre_processing:

Severity records for Administrative and systemic events will be filtered
from the input data set and AVAL will be derived in a data_frame pp

Maximum severity derivation:

Once the pre-processing is done. The maximum severity records will be derived
from the  AVAL by USUBJID and by_vars.
Below variables will be created with the values mentioned below.
'DTYPE = "MAXIMUM", FATESTCD = "MAXSEV" and FATEST = "MAXIMUM SEVERITY".'
}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_maxtemp}()},
\code{\link{hello_admiral}()}
}
\concept{der_adxx}
\keyword{der_adxx}
