% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_maxtemp.R
\name{derive_param_maxtemp}
\alias{derive_param_maxtemp}
\title{Deriving maximum temperature record}
\usage{
derive_param_maxtemp(
  dataset = NULL,
  filter_faobj = "FEVER",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF),
  test_maxtemp = "Maximum temperatue",
  testcd_maxtemp = "MAXTEMP"
)
}
\arguments{
\item{dataset}{Input data set

Input data set should have the temperature record which will be derived by
using the function \code{derive_param_fever_occur}.}

\item{filter_faobj}{filter for fever record

\emph{Default: "FEVER"}
\emph{Permitted Value: A Character Scalar}

Helps to filter the fever record by passing the value from \code{FATESTCD} which
indicates the fever records.}

\item{by_vars}{Grouping variables

\emph{Default: vars(USUBJID,FAOBJ,ATPTREF)}
\emph{Permitted Value: A Character vector}

Pass the \code{by_vars} in \code{vars()} based on the by_vars variables the maximum
temperature will be derived from \code{VSSTRESN}.}

\item{test_maxtemp}{Value for \code{FATEST}

\emph{Default}: "Maximum temperature"
\emph{Permitted Value: A Character scalar}

Pass a value for \code{FATEST} which will be indicate the maximum temperature
records}

\item{testcd_maxtemp}{\emph{Default}: "MAXTEMP"
\emph{Permitted Value: A Character scalar}

Pass a value for \code{FATESTCD} which will be indicate the maximum temperature
records. refer the default value.}
}
\value{
The input data set with derived records indicating the maximum temperature
records.
}
\description{
Create a summary record for maximum temperature per subject per period.
}
\examples{
library(tibble)
library(admiral)
library(dplyr)
library(rlang)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~ATPTREF, ~AVALC, ~ATPTN, ~VSSTRESN, ~FATEST, ~FATESTCD,
  "XYZ1001", "FEVER", "VACC 1", "Y", 1, 38.9, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 1", "Y", 2, 39.2, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 1", "Y", 3, 39.2, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 1", "Y", 4, 39.2, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 1", "N", 5, 33.1, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 1", "Y", 6, 36.0, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 2", "Y", 1, 38.9, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 2", "Y", 2, 37.2, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 2", "Y", 3, 37.3, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 2", "Y", 4, 36.2, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 2", "N", 5, 36.1, "Occurrence Indicator", "OCCUR",
  "XYZ1001", "FEVER", "VACC 2", "Y", 6, 34.0, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 1", "N", 1, 36.7, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 1", "Y", 2, 38.2, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 1", "Y", 3, 35.5, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 1", "Y", 4, 38.2, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 1", "Y", 5, 37.5, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 1", "Y", 6, 37.2, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 2", "Y", 1, 38.0, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 2", "N", 2, 37.2, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 2", "N", 3, 37.4, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 2", "Y", 4, 38.2, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 2", "N", 5, 36.1, "Occurrence Indicator", "OCCUR",
  "XYZ1002", "FEVER", "VACC 2", "N", 6, 36.2, "Occurrence Indicator", "OCCUR"
)

derive_param_maxtemp(
  dataset = input,
  filter_faobj = "FEVER",
  test_maxtemp = "Maximum Temperature",
  testcd_maxtemp = "MAXTEMP",
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF)
)

}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\concept{der_adxx}
\keyword{der_adxx}
