% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_crit.R
\name{derive_vars_crit}
\alias{derive_vars_crit}
\title{Derive Analysis Criterion Evaluation variables}
\usage{
derive_vars_crit(dataset, new_var, label_var, condition, criterion)
}
\arguments{
\item{dataset}{Input dataset}

\item{new_var}{Variables to add
The analysis criterion evaluation variable's name (i.e., CRIT1)
This name is also used in order to create both charcater and numeric
flags variables (i.e., CRIT1FL and CRIT1FN).
Is possible to give a different name, base on needs (i.e., ANL01)}

\item{label_var}{Criterion value
A text description defining the condition necessary to satisfy the presence
of the criterion}

\item{condition}{Condition for selecting a subset
The condition specified in order to select a subset from the input dataset
in which the rule is applied.}

\item{criterion}{Criterion rule
The criterion that each selected row satisfies or not.
Returns Y or N for character variable and 1 or 0 fro numeric variable
if the criterion is met or not, respectively.
Returns NA for not selected rows (not taken into account from condition)}
}
\value{
dataset with criterion variables
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                                                              +
}\if{html}{\out{</div>}}

Derive analysis criterion evaluation result variable, paired with character    +
and numeric flags.                                                             +
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
}
\details{
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
}
\examples{

is2 <- derive_vars_crit(
  dataset = is,
  new_var = "CRIT1",
  label_var = "Titer >= ISLLOQ",
  condition = !is.na(AVAL) & !is.na(ISLLOQ),
  criterion = AVAL >= ISLLOQ
)


}
\author{
Federico Baratin
}
