% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_crit.R
\name{derive_vars_crit}
\alias{derive_vars_crit}
\title{Derive Analysis Criterion Evaluation Variables}
\usage{
derive_vars_crit(dataset, prefix, crit_label, condition, criterion)
}
\arguments{
\item{dataset}{Input dataset}

\item{prefix}{Variables to add

The analysis criterion evaluation variable's name (i.e., \code{CRIT1})
This name is also used in order to create both character and numeric
flags variables (i.e., \code{CRIT1FL} and \code{CRIT1FN}).
If the name does not contain CRIT wording, it generates a flag variable
(ex: \code{ANL01FL}) whose logic is equals to \code{CRIT1} variable, without generating
additional numeric (ex: \code{ANL01FN}) and character (\code{ANL01}) variables.}

\item{crit_label}{Criterion value

A text description defining the condition necessary to satisfy the presence
of the criterion}

\item{condition}{Condition for selecting a subset

The condition specified in order to select a subset from the input dataset
in which the rule is applied.}

\item{criterion}{Criterion rule

The criterion that each selected row satisfies or not.
Returns \code{Y} or \code{N} for character variable and \code{1} or \code{0} for numeric variable
if the criterion is met or not, respectively.
Returns \code{NA} for not selected rows (not taken into account from condition)}
}
\value{
Dataset with criterion variables
}
\description{
\verb{r derive_vars_crit::admiralvaccine("deprecated")}

This function is \emph{deprecated}, please use \code{derive_vars_crit_flag} instead.
}
\author{
Federico Baratin
}
\concept{deprecated}
\keyword{deprecated}
