% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_event_flag.R
\name{derive_vars_event_flag}
\alias{derive_vars_event_flag}
\title{Adds Flag Variables for an Occurred Event .}
\usage{
derive_vars_event_flag(
  dataset,
  by_vars,
  aval_cutoff,
  new_var1 = NULL,
  new_var2 = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are
expected.}

\item{by_vars}{Grouping variables

\if{html}{\out{<div class="sourceCode">}}\preformatted{  *Permitted values:* list of variables created by `exprs()`
}\if{html}{\out{</div>}}

The variables to be considered for grouping for creating a new
variable \code{new_var1}}

\item{aval_cutoff}{Cutoff value

The cutoff value to decide on whether to consider the event has occurred
or not for the \code{TESTCD} based on diameter. For example, if the diameter
value is greater than x only we will consider the event occurred for
that record otherwise it will be considered has no event happened for
that record.}

\item{new_var1}{Name of the new flag variable 1

\if{html}{\out{<div class="sourceCode">}}\preformatted{A new flag variable will be created with values "Y" or "N".
If the event is occurred at least once during a observation period for
a subject then the new variable will be flagged as Y otherwise N.
}\if{html}{\out{</div>}}}

\item{new_var2}{Name of the new flag variable 2.

\if{html}{\out{<div class="sourceCode">}}\preformatted{A new flag variable will be created with values "Y" or "N".
If the event is occurred on the particular day then the new variable will
be flagged as Y otherwise N.
}\if{html}{\out{</div>}}}
}
\value{
the dataset with the flag variables added to it.
}
\description{
Creates two flag variables for the event occurred,one for the event occurred
within each by group and one to flag if the event occurred or not
for each day.
}
\details{
The names for the new flag variables created will be sponsor specific.

For the \code{new_var1} it will flag all observations as "Y" within the by group
if the event occurred at least once during observation period. If the event
is not at all occurred during the observation period then all the
observations within by group will be flagged as "N".

For derived records with \code{DTYPE} equals "MAXIMUM" , the \code{new_var2} will be
set to NA.
}
\examples{
library(tibble)
library(admiral)
library(dplyr)
library(rlang)

input <- tribble(
  ~USUBJID, ~FAOBJ, ~ATPTREF, ~AVAL, ~AVALC, ~FATEST, ~FATESTCD, ~FASCAT, ~DTYPE,
  "1", "REDNESS", "VAC1", 3.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE", "",
  "1", "REDNESS", "VAC1", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE", "",
  "1", "REDNESS", "VAC1", 1.5, "3.5", "Diameter", "DIAMETER", "ADMIN-SITE", "",
  "1", "REDNESS", "VAC1", 4.5, "4.5", "Diameter", "DIAMETER", "ADMIN-SITE", "MAXIMUM",
  "1", "FATIGUE", "VAC1", 1, "MILD", "Severity", "SEV", "SYSTEMIC", "",
  "1", "FATIGUE", "VAC1", 2, "MODERATE", "Severity", "SEV", "SYSTEMIC", "",
  "1", "FATIGUE", "VAC1", 0, "NONE", "Severity", "SEV", "SYSTEMIC", "",
  "1", "FATIGUE", "VAC1", 2, "MODERATE", "Severity", "SEV", "SYSTEMIC", "MAXIMUM",
  "1", "REDNESS", "VAC2", 6.5, "6.5", "Diameter", "DIAMETER", "ADMIN-SITE", "",
  "1", "REDNESS", "VAC2", 7.5, "7.5", "Diameter", "DIAMETER", "ADMIN-SITE", "",
  "1", "REDNESS", "VAC2", 2.5, "2.5", "Diameter", "DIAMETER", "ADMIN-SITE", "",
  "1", "REDNESS", "VAC2", 7.5, "7.5", "Diameter", "DIAMETER", "ADMIN-SITE", "MAXIMUM",
  "1", "FATIGUE", "VAC2", 1, "MILD", "Severity", "SEV", "SYSTEMIC", "",
  "1", "FATIGUE", "VAC2", 2, "MODERATE", "Severity", "SEV", "SYSTEMIC", "",
  "1", "FATIGUE", "VAC2", 0, "NONE", "Severity", "SEV", "SYSTEMIC", "",
  "1", "FATIGUE", "VAC2", 2, "MODERATE", "Severity", "SEV", "SYSTEMIC", "MAXIMUM",
)

derive_vars_event_flag(
  dataset = input,
  by_vars = exprs(USUBJID, FAOBJ, ATPTREF),
  aval_cutoff = 2.5,
  new_var1 = EVENTL,
  new_var2 = EVENTDL
)

}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\concept{der_adxx}
\keyword{der_adxx}
