% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_max_flag.R
\name{derive_vars_max_flag}
\alias{derive_vars_max_flag}
\title{\code{derive_vars_max_flag.R}}
\usage{
derive_vars_max_flag(dataset, flag1 = "ANL01FL", flag2 = "ANL02FL")
}
\arguments{
\item{`dataset`}{Input dataset}

\item{`flag1`}{\itemize{
\item Flags the maximum record per subject per event per
Vaccination
\emph{Default: "ANL01FL"}
\emph{Permitted value: Any variable name or NULL}
\code{NULL} denotes not to create the flag
}}

\item{`flag2`}{\itemize{
\item Flags the maximum record per subject per event for Overall
\emph{Default: "ANL02FL"}
\emph{Permitted value: Any variable name or NULL}
\code{NULL} denotes not to create the flag
}}
}
\value{
The output dataset creates \code{ANLxxFL} flags
}
\description{
\code{derive_vars_max_flag.R}
}
\details{
This utility flags the maximum record per subject per event
per vaccination/Overall
If both parameters \code{flag1} & \code{flag2} are passed as NULL then
utility will throw error and flags will not be created.
}
\examples{
library(dplyr)
library(tidyverse)
library(admiraldev)
library(admiral)
library(tibble)
input <- tribble(
  ~USUBJID, ~FAOBJ, ~FATESTCD, ~FATPTREF, ~AVAL, ~FADTC, ~PARAMCD,
  "ABC101", "REDNESS", "DIAMETER", "VACC 1", 10, "2015-01-10", "DIARE",
  "ABC101", "REDNESS", "DIAMETER", "VACC 1", 7, "2015-01-11", "DIARE",
  "ABC101", "REDNESS", "DIAMETER", "VACC 2", 3, "2015-02-10", "DIARE",
  "ABC101", "REDNESS", "DIAMETER", "VACC 2", 8, "2015-02-11", "DIARE",
  "ABC101", "FATIQUE", "SEV", "VACC 1", 1, "2015-01-10", "SEVFAT",
  "ABC101", "FATIQUE", "SEV", "VACC 1", 1, "2015-01-11", "SEVFAT",
  "ABC101", "FATIQUE", "SEV", "VACC 2", 2, "2015-02-10", "SEVFAT",
  "ABC101", "FATIQUE", "SEV", "VACC 2", 3, "2015-02-11", "SEVFAT"
)

derive_vars_max_flag(
  dataset = input,
  flag1 = "ANL01FL",
  flag2 = "ANL02FL"
)
derive_vars_max_flag(
  dataset = input,
  flag1 = NULL,
  flag2 = "ANL02FL"
)
derive_vars_max_flag(
  dataset = input,
  flag1 = "ANL01FL",
  flag2 = NULL
)

}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\author{
Dhivya Kanagaraj
}
\concept{der_adxx}
\keyword{der_adxx}
