% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_merged_vaccine.R
\name{derive_vars_merged_vaccine}
\alias{derive_vars_merged_vaccine}
\title{Add New Variable(s) to the Input dataset Based on Variables from Another dataset}
\usage{
derive_vars_merged_vaccine(
  dataset,
  dataset_ex,
  by_vars_sys,
  by_vars_adms,
  dataset_supp = NULL,
  dataset_suppex = NULL,
  ex_vars
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument inside the
\code{derive_vars_merged}are expected.}

\item{dataset_ex}{EX dataset to merge with the input dataset.

The variables specified by the \code{ex_vars} argument are expected.}

\item{by_vars_sys}{Grouping variables for systemic events.

\emph{Default: exprs(USUBJID, FATPTREF)}}

\item{by_vars_adms}{Grouping variables for administration site events.

\emph{Default: exprs(USUBJID, FATPTREF, FALOC, FALAT, FADIR)}}

\item{dataset_supp}{Supplementary input dataset
By default \code{dataset_supp} will be NULL,user has to provide
supplementary dataset to merge it back with original input dataset
if they have supplementary dataset in their case.}

\item{dataset_suppex}{Supplementary EX dataset
By default \code{dataset_suppex} will be NULL,user has to provide
supplementary dataset to merge it back with original EX dataset
if they have supplementary dataset in their case.}

\item{ex_vars}{Variables to be added to the output dataset from EX dataset}
}
\value{
the dataset with variables added from the EX dataset.
}
\description{
Add New Variable(s) to the Input dataset Based on Variables from Another dataset.
The variables to be added to the output dataset will be based on input variables
passed on \code{ex_vars} argument
}
\details{
The input dataset will be merged with Ex dataset for "ADMINISTRATION SITE" and
"SYSTEMIC" categories separately and these datasets will be binded together as
the final output dataset.

This function is intended to add only \code{EX} variables to the input dataset, an user
is expected to handle if any pre-processing is required.

Only the variables passed to the \code{ex_vars} will be present in the output dataset

If the input dataset has multiple vaccination for a subject at same visit
then this function will not merge ex dataset and will return only the input
dataset merged with its supplementary dataset.
}
\examples{

library(tibble)
library(admiral)
library(dplyr)
library(rlang)

derive_vars_merged_vaccine(
  dataset = vx_face,
  dataset_ex = vx_ex,
  dataset_supp = NULL,
  dataset_suppex = NULL,
  by_vars_sys = exprs(USUBJID, FATPTREF),
  by_vars_adms = exprs(USUBJID, FATPTREF, FALOC, FALAT, FADIR),
  ex_vars = exprs(EXTRT, EXDOSE, EXDOSU, EXSTDTC, EXENDTC)
)

}
\seealso{
Other der_var: 
\code{\link{derive_vars_crit}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()}
}
\author{
Vikram S
}
\concept{der_var}
\keyword{der_var}
