% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_params.R
\name{derive_vars_params}
\alias{derive_vars_params}
\title{\code{derive_vars_params.R}}
\usage{
derive_vars_params(dataset, lookup_dataset)
}
\arguments{
\item{dataset}{\itemize{
\item Input dataset
Input dataset is expected to have variables \code{USUBJID},\code{FAOBJ},
\code{FACAT}, \code{FATESTCD} and \code{FATEST}
}}

\item{lookup_dataset}{\itemize{
\item lookup file containing \code{PARAMCD} values for every
unique \code{FATESTCD} and \code{FAOBJ}
lookup file is expected to have variables \code{FATEST}, \code{PARAMCD},
\code{FATESTCD}, \code{FAOBJ} and one entry for every unique
\code{FATESTCD} and \code{FAOBJ}
}}
}
\value{
The output dataset contains all observations and variables of the
input dataset along with \code{PARAM},\code{PARAMCD},\code{PARCAT1},\code{PARCAT2},\code{PARAMN}
}
\description{
Creating \code{PARAMCD} from lookup file and assigning \code{PARAM},\code{PARAMN},\code{PARCAT1},
\code{PARCAT2} variables
}
\details{
A lookup file is required with PARAMCD values for every combination
of \code{FATEST} & \code{FAOBJ}.
\code{PARAMCD} values comes from lookup file.
\code{PARAMN} is assigned with a unique number for every unique PARAM value.
\code{PARAM} value is a combination of \code{FAOBJ} \code{FATEST} \code{FASTRESU} \code{FALOC}
\code{FADIR} \code{FALAT}
}
\examples{

library(tibble)
lookup_dataset <- tribble(
  ~FATESTCD,    ~PARAMCD,    ~FATEST,                ~FAOBJ,
  "SEV",        "SEVREDN",   "Severity",             "Redness",
  "DIAMETER",   "DIARE",     "Diameter",             "Redness",
  "MAXDIAM",    "MDIRE",     "Maximum Diameter cm",  "Redness",
  "MAXTEMP",    "MAXTEMP",   "Maximum Temperature",  "Fever",
  "OCCUR",      "OCFEVER",   "Occurrence Indicator", "Fever",
  "OCCUR",      "OCERYTH",   "Occurrence Indicator", "Erythema",
  "SEV",        "SEVPAIN",   "Severity",             "Pain at Injection site",
  "OCCUR",      "OCPAIN",    "Occurrence Indicator", "Pain at Injection site",
  "OCCUR",      "OCSWEL",    "Occurrence Indicator", "Swelling"
)

input <- tribble(
  ~USUBJID, ~FACAT, ~FASCAT, ~FATESTCD, ~FAOBJ, ~FATEST, ~FALOC, ~FALAT,
  "ABC101", "REACTO", "ADMIN", "SEV", "Redness", "Severity", "ARM", "LEFT",
  "ABC101", "REACTO", "ADMIN", "DIAMETER", "Redness", "Diameter", "ARM", "RIGHT",
  "ABC101", "REACTO", "ADMIN", "MAXDIAM", "Redness", "Maximum Diameter", NA, NA,
  "ABC101", "REACTO", "SYSTEMIC", "MAXTEMP", "Fever", "Maximum Temp", NA, NA,
  "ABC101", "REACTO", "SYSTEMIC", "OCCUR", "Fever", "Occurrence", NA, NA,
  "ABC101", "REACTO", "ADMIN", "OCCUR", "Erythema", "Occurrence", NA, NA,
  "ABC101", "REACTO", "ADMIN", "SEV", "Swelling", "Severity", NA, NA,
  "ABC101", "REACTO", "ADMIN", "OCCUR", "Swelling", "Occurrence", NA, NA,
  "ABC101", "REACTO", "ADMIN", "OCCUR", "Swelling", "Occurrence", NA, NA
)
derive_vars_params(
  dataset = input,
  lookup_dataset = lookup_dataset
)

}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_vaxdt}()},
\code{\link{hello_admiral}()}
}
\author{
Dhivya Kanagaraj
}
\concept{der_adxx}
\keyword{der_adxx}
