% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_vaxdt.R
\name{derive_vars_vaxdt}
\alias{derive_vars_vaxdt}
\title{Adds vaccination date variables to the output dataset.}
\usage{
derive_vars_vaxdt(dataset, dataset_adsl, by_vars, order)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected.}

\item{by_vars}{Grouping variables.

\emph{Default: exprs(USUBJID, VISITNUM)}

The variables to be grouped to filter the first observation within each
by group.}

\item{order}{Sorting variables.

\emph{Default: exprs(USUBJID,VISITNUM,VISIT,EXSTDTC)}

The variables order to be specified either in ascending or descending order.
By default ascending order will be applicable.}
}
\value{
the dataset with vaccination date variables which is created
using \code{EXSTDTC}.
}
\description{
Creates vaccination date variables from ex domain.A date variable will be
created for each vaccination taking values from the variable \code{EXSTDTC}.
}
\details{
If there are multiple vaccinations for a visit per subject,only first
observation will be filtered based on the variable order specified
on the \code{order} argument.

The number of variables created will be based on the number of vaccinations
per subject per visit.

The output dataset will have one record per subject.
}
\examples{
library(tibble)
library(admiral)
library(dplyr)
library(rlang)

input <- tribble(
  ~USUBJID, ~EXSTDTC, ~VISITNUM, ~EXTRT, ~EXLNKGRP, ~VISIT,
  "A001", "2015-01-10", 1, "DRUG A", "VAC 1", "VISIT 1",
  "A001", "2015-01-11", 2, "DRUG A", "VAC 2", "VISIT 2",
  "A001", "2015-01-12", 3, "DRUG B", "VAC 3", "VISIT 3",
  "A002", "2015-01-13", 1, "DRUG B", "VAC 1", "VISIT 1",
  "A002", "2015-01-14", 2, "DRUG C", "VAC 2", "VISIT 2"
)

adsl <- tribble(
  ~USUBJID, ~SEX, ~AGE,
  "ABC001", "MALE", 23,
  "ABC002", "FEMALE", 26,
)
}
\seealso{
ADXX Functions that returns variable appended to dataset: 
\code{\link{derive_param_diam_to_sev}()},
\code{\link{derive_param_fever_occur}()},
\code{\link{derive_param_maxdiam}()},
\code{\link{derive_param_maxsev}()},
\code{\link{derive_param_maxtemp}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{hello_admiral}()}
}
\author{
Vikram S
}
\concept{der_adxx}
\keyword{der_adxx}
