% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_max_flag.R
\name{max_flag}
\alias{max_flag}
\title{Creating Maximum Flag}
\usage{
max_flag(dataset, by_vars, by_join, fl)
}
\arguments{
\item{dataset}{Input dataset}

\item{by_vars}{By variables which goes to group by, to create the flag. Pass the variables it
inside the exprs().}

\item{by_join}{By variables to join the flagged records with input dataset.Pass the variables in
a vector.}

\item{fl}{Flag variable name, Pass it as string.}
}
\value{
data frame with flag variable which is flagged for the maximum value records depends on
the variables passed in \code{by_vars} by user.
}
\description{
To Flag the maximum records depends on the grouping varibales in a flag varibale.
}
\examples{
input <- tribble(
  ~USUBJID, ~FAOBJ, ~FATESTCD, ~FATPTREF, ~AVAL, ~FADTC, ~PARAMCD,
  "ABC101", "REDNESS", "DIAMETER", "VACC 1", 10, "2015-01-10", "DIARE",
  "ABC101", "REDNESS", "DIAMETER", "VACC 1", 7, "2015-01-11", "DIARE",
  "ABC101", "REDNESS", "DIAMETER", "VACC 2", 3, "2015-02-10", "DIARE",
  "ABC101", "REDNESS", "DIAMETER", "VACC 2", 8, "2015-02-11", "DIARE",
  "ABC101", "FATIQUE", "SEV", "VACC 1", 1, "2015-01-10", "SEVFAT",
  "ABC101", "FATIQUE", "SEV", "VACC 1", 1, "2015-01-11", "SEVFAT",
  "ABC101", "FATIQUE", "SEV", "VACC 2", 2, "2015-02-10", "SEVFAT",
  "ABC101", "FATIQUE", "SEV", "VACC 2", 3, "2015-02-11", "SEVFAT"
)

max_flag(
  dataset = input,
  by_vars = exprs(USUBJID, FAOBJ, FATPTREF, PARAMCD),
  by_join = c("USUBJID", "FAOBJ", "FATPTREF", "PARAMCD", "FADTC"),
  fl = "ANL01FL"
)
}
\seealso{
Other der_var: 
\code{\link{derive_vars_crit}()},
\code{\link{derive_vars_event_flag}()},
\code{\link{derive_vars_max_flag}()},
\code{\link{derive_vars_merged_vaccine}()},
\code{\link{derive_vars_params}()},
\code{\link{derive_vars_vaxdt}()}
}
\author{
Dhivya Kanagaraj
}
\concept{der_var}
\keyword{der_var}
