% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppis}
\alias{suppis}
\title{Supplemental IS Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 10 columns.
}
\usage{
suppis
}
\description{
A SDTM CM dataset
}
\keyword{datasets}
