% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_adsl}
\alias{vx_adsl}
\title{Vx - Subject Level Analysis Dataset}
\format{
An object of class \code{data.frame} with 10 rows and 14 columns.
}
\usage{
vx_adsl
}
\description{
A subject level analysis dataset
}
\seealso{
Other dataset: 
\code{\link{adsl}},
\code{\link{ae}},
\code{\link{ce}},
\code{\link{dm}},
\code{\link{ex}},
\code{\link{face}},
\code{\link{is}},
\code{\link{suppis}},
\code{\link{vs}}
}
\concept{dataset}
\keyword{dataset}
