% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_adsl}
\alias{vx_adsl}
\title{Subject Level Analysis - Vaccine Specific}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 47 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_adsl.rda}
}
\usage{
vx_adsl
}
\description{
A subject level analysis dataset
}
\seealso{
Other dataset: 
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppce}},
\code{\link{vx_suppdm}},
\code{\link{vx_suppex}},
\code{\link{vx_suppface}},
\code{\link{vx_suppis}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
