% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_dm}
\alias{vx_dm}
\title{DM Dataset}
\format{
An object of class \code{data.frame} with 2 rows and 28 columns.
}
\usage{
vx_dm
}
\description{
A SDTM DM dataset
}
\seealso{
Other dataset: 
\code{\link{adface}},
\code{\link{vx_adsl}},
\code{\link{vx_ae}},
\code{\link{vx_ce}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppis}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
