% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_face}
\alias{vx_face}
\title{Findings About Clinical Events}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 307 rows and 30 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_face.rda}
}
\usage{
vx_face
}
\description{
A SDTM FACE dataset
}
\seealso{
Other dataset: 
\code{\link{vx_adsl}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_is}},
\code{\link{vx_suppce}},
\code{\link{vx_suppdm}},
\code{\link{vx_suppex}},
\code{\link{vx_suppface}},
\code{\link{vx_suppis}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
