% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_suppis}
\alias{vx_suppis}
\title{Supplemental Immunogenicity Specimen}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 10 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/inst/create_vx_data/create_suppis.R}
}
\usage{
vx_suppis
}
\description{
A SDTM Supplementary IS dataset
}
\seealso{
Other dataset: 
\code{\link{vx_adsl}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppce}},
\code{\link{vx_suppdm}},
\code{\link{vx_suppex}},
\code{\link{vx_suppface}},
\code{\link{vx_vs}}
}
\concept{dataset}
\keyword{dataset}
