% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vx_vs}
\alias{vx_vs}
\title{Vital Signs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 23 columns.
}
\source{
\url{https://github.com/pharmaverse/admiralvaccine/blob/main/data/vx_vs.rda}
}
\usage{
vx_vs
}
\description{
A SDTM VS dataset
}
\seealso{
Other dataset: 
\code{\link{vx_adsl}},
\code{\link{vx_ce}},
\code{\link{vx_dm}},
\code{\link{vx_ex}},
\code{\link{vx_face}},
\code{\link{vx_is}},
\code{\link{vx_suppce}},
\code{\link{vx_suppdm}},
\code{\link{vx_suppex}},
\code{\link{vx_suppface}},
\code{\link{vx_suppis}}
}
\concept{dataset}
\keyword{dataset}
